Issue #11163: Split PackageName input file to enforce 120-line limit
